requirements.txt
Django==4.0
django-debug-toolbar==3.2.2
django-extensions==3.1.5
django-allauth==0.46.0
django-htmx==1.6.0
django-crispy-forms==1.13.0
crispy-bootstrap5==0.6

psycopg2-binary==2.9.2
dj-database-url==0.5.0
python-decouple==3.5
python-dateutil==1.16.0
pandas-datareader==0.10.0
pytz==2021.3
asgiref==3.4.1

gunicorn==20.1.0
